git-svn-id: https://gorgonlib.googlecode.com/svn/branches/2.x@1454 a61f9e97-bd4c-0410-8c65-85b29eaa70ff

+++ Dorian/Gorgon/Gorgon.Graphics/Fonts/GorgonGlyph.cs
@@ -135,10 +135,10 @@ namespace GorgonLibrary.Graphics
 
 			Character = character;
 			GlyphCoordinates = glyphCoordinates;
-			TextureCoordinates = RectangleF.FromLTRB((float)glyphCoordinates.Left / (float)texture.Settings.Width,
-												(float)glyphCoordinates.Top / (float)texture.Settings.Height,
-												(float)glyphCoordinates.Right / (float)texture.Settings.Width,
-												(float)glyphCoordinates.Bottom / (float)texture.Settings.Height);
+			TextureCoordinates = RectangleF.FromLTRB(glyphCoordinates.Left / (float)texture.Settings.Width,
+												glyphCoordinates.Top / (float)texture.Settings.Height,
+												glyphCoordinates.Right / (float)texture.Settings.Width,
+												glyphCoordinates.Bottom / (float)texture.Settings.Height);
 			Texture = texture;
 			Offset = glyphOffset;
 			Advance = glyphAdvancing;
